core:nw:Add method to get content of request header
There is no way to get access to content of a request header unless
header content is of type "ContentType::FormUrlEncoded". So added a new
method HttpRequest::getHeaderData() to return header content as a String.

So if you are posing data to HTTP server with mime type different than
"application/x-www-form-urlencoded", you can use getHeaderData in
http request callback function for parsing.

Signed-off-by: Ajay Bhargav <contact@rickeyworld.info>