Remove '-u user_init' from LDFLAGS
This was intended to force the use of the local compiled user_init instead of the pre-compiled version in libsming. However, it causes some odd other problems which I cannot explain. Until there is a fix for #294 to ensure the right user_init is pulled in (and without causing these odd problems) you should compile libsming with the spiffs_mount call commented out in Sming/Sming/appinit/user_main.cpp when you wish to use rBoot roms.