Wait for header parsing to writeRawData
Some web servers (python 2.7 SimpleHTTPServer, at least), send the HTTP/1.1 200 OK line as one packet, the headers in the next packet, and the body starting in the third packet.  Without this change, writeRawData will write the headers on these servers, but start with the body on servers that send the status code and headers together in one packet.  By waiting until the headerEnd is confirmed, this makes the HTTP client act consistently across all servers.