Update serial examples to latest changes in printf.
Using the optional printf flag '#' (as in "%#x") in the examples is
some what misleading.

The old code simply ignores the flag and always adds prefix "0x"
regardless, which is against the C printf standard (and looks pretty
ugly in some cases).

The new code also ignores the flag, but does not add the 0x prefix

The '#' example is left there,  with a note. That hopefully makes it
serve as a feature (a reminder of a minor missing feature) rather
than a bug.

Signed-off-by: Quy Tonthat <qtonthat@gmail.com>