use COM_SPEED_SERIAL from Makefile instead of SERIAL_BAUD_RATE
the Makefiles define COM_SPEED_SERIAL and use this to start the
terminal, but this value isn't passed to the code. inside user_config.h
there is SERIAL_BAUD_RATE that configures the port speed. this means
it's left up to the user to keep the two in sync, which is less than
ideal. this commit passes COM_SPEED_SERIAL to gcc as a define, and
changes all Serial.begin lines to use this instead of SERIAL_BAUD_RATE.