make COM_SPEED_SERIAL from Makefile available inside apps
the Makefiles define COM_SPEED_SERIAL and uses this to start the
terminal, but this value isn't passed to the code. inside user_config.h
there is SERIAL_BAUD_RATE that configures the port speed. this means
it's left up to the user to keep the two in sync, which is less than
ideal. this commit passes COM_SPEED_SERIAL to gcc as a define, and
allows users to use one value for both if they chooseto do so. this
won't break any existing code.