Initial refactoring of the WebsocketClient.
Renamed the method making the `binding` to a http server socket to be called `bind`.

Fix for potential memory leak + Codacy fixes.

Reapplied coding-standards.

Small changes.

Updated the example to update every 20 seconds instead of 60.
Better init and deinit of streams

Added timeouts to keep the connection for longer periods.

Reset the HttpConnection timeout on a new connect.

- Don't send new HTTP request over the same TCP connection if is already started.
- Make HttpConnection::reset to be extendable. This allows also the WebsocketClient to reset the connection.

Use SSL by default

Refactored the HttpClient and Connection to reuse code.
Refactored the WebsocketConnection to be reused in server and client mode.