Create SimpleTimer class to wrap hardware and revise Timer (#1477)
There are many instances where the extended capabilities of Timer are not required. Complex applications can use a fair number of these, each of which consumes more RAM than necessary for the task it is required to perform. The developer may therefore choose to 'downgrade' to an OSTimer to recover the RAM.