Refactoring websocket client and server. (#1514)
## Improvements
* Simplified and decreased the code base responsible for the Websocket functionality. 
Refactored the WebsocketConnection to be reused in server and client mode.

* Documentation for the updated connections.

## Breaking Changes (Impact: High)

If you use Websocket functionality make sure to read the notes below. 

* Renamed Websocket related types, classes and methods. We used to have `WebSocket` and `Websocket`.  From now on only `Websocket` will be used. Example ` WebsocketConnection::getActiveWebSockets` is now ` WebsocketConnection::getActiveWebsockets`.

* WebsocketConnection
Renamed the method making the `binding` to a http server socket to be called `bind`.

* WebsocketClient
All methods starting with `setWebSocket` are shortened to `set`. Example: `setWebSocketMessageHandler` is renamed to `setMessageHandler`.