Added testing OTA server (#1537)
It is now possible to call
```
make otaserver
```
and start a web server that can be used to TEST OTA updates. 

The default port is 9999. If you want to use another port make sure to specify it from the command line. Example for port 8080 is given below:
```
make otaserver SERVER_OTA_PORT=8080
```