Change use of `ReadWriteStream*` to `IDataSourceStream*` where appropriate (#1588)
* Change use of `ReadWriteStream*` to `IDataSourceStream*` where appropriate

Changes to parameter types, member variables
`MultiStream`and `StreamTransformer` classes now inherit from `IDataSourceStream` instead of `ReadWriteStream`; this affects all inherited classes

With `TcpClient`, a stream is used for outgoing data so should be inherited from IDataSourceStream, not ReadWriteStream. An additional internal 'buffer' ReadWriteStream is used where necessary.
This is also done for the `HttpResponse`class.