Add optional `deferMillis` parameter to `System::restart()` (#1611)
A system restart may be requested from within a deeply-nested callback routine, such as on network transfer completion. This happens in the `Basic_rBoot` sample application, which I have found occasionally crashes at the end of an update cycle; this is probably more likely in debug mode where there is more going on.

In such scenarios it is best to defer the actual restart request so that the callback can unwind and complete fully. Therefore, this modification will _always_ defer a restart request, using either the task queue (if deferMillis == 0) or a system timer if a longer delay is required by the application.