Fix exception related to `ENABLE_CMD_EXECUTOR` definition
If framework is built with `ENABLE_CMD_EXECUTOR` disabled (0), but application is built with it enabled, serial receive causes an exception.
This happens because the member data for `HardwareSerial` gets shifted by 4 bytes, so `uart` becomes invalid (we get `options` instead, usually 1).

It is safer to always build with `HardwareSerial::commandExecutor` included.