Fix issue with debugger break during syscall
When a syscall is being processed, especially a lengthy one like reading from the console, a regular debug break or exception takes us into the debugger command loop but GDB doesn't know we've stopped (it ignores our 'T' message). Then, when the 'F' response (which completes the syscall) is received, it takes us back out of the debugger loop immediately, which is wrong; at that point we should tell GDB again why we've stopped so things can proceed as for a regular break handling.

ONGOING: When testing with a breakpoint in a regular system timer the stub hangs during the `gdbstub_icount_ena_single_step()` call - this has been temporarily highlighted with the !! debug statements.