Remove `UNAME` from output build path, not really needed
This changes the output paths from:

`out/build/Host/Windows/debug/`

to just

`out/build/Host/debug/`

This was added in the first place to allow the same working tree to be used for both Windows and Linux Host builds to simplify testing. For instance, running Linux in a VM on a Windows machine.

In retrospect it works better if I just use the Windows working tree as the main repo, and clone it in Linux.