Remove `UNAME` from output build path, not really needed (#1769)
This changes the output paths from:

`out/build/Host/Windows/debug/`

to just

`out/build/Host/debug/`

This was added in the first place to allow the same working tree to be used for both Windows and Linux Host builds to simplify testing. For instance, running Linux in a VM on a Windows machine.