Define `-fno-threadsafe-statics` for Esp8266
Static variables are initialised on first use and may result in additional calls to `__cxa_guard_acquire()` and `__cxa_guard_release()`.
If statics are contained within an interrupt service routine it may cause it to fail as the guard functions aren't in IRAM.

This isn't necessary and can be disabled using the above compiler flag.