Revise UART driver and HardwareSerial to avoid using `peri.h` (#1836)
* Revise UART driver and HardwareSerial to avoid using `peri.h`.

Definitions in peri.h are too succinct and hard to read. Instead, use the `uart_register.h` definitions from the RTOS SDK.

Note: The 'struct' register approach is clearer, but requires more extensive changes to the driver.

Decouple `SerialStatus` flag bits from driver. Only a few definitions so just map them.

* Fix printf format strings in Basic_Serial sample

* Use large TX buffers and no wait in Basic_Serial to better test asynchronous operation

* Fix `ReadWriteStream::copyFrom` method to check write count.

This is the intended behaviour, got missed when this method was added in #1701.