Add `ENABLE_TASK_COUNT` to control task counter, disable by default.
When stressed with periodic interrupts, the only way to get a reliable count is to disable interrupts
whilst updating the task counts. As this is a diagnostic feature, leaving it disabled
by default seems appropriate.

Fortunately, `noInterrupts()` already returns the existing interrupt level,
so we just need to add `restoreInterrupts()` so we can safely use it from interrupt context.