Address inconsistencies when setting `DATE` field in HTTP headers.
[RFC 2616](https://tools.ietf.org/html/rfc2616#section-14.18) can probably be summed up as:
*If you can provide a valid DATE, then do so*.

Which begs the question, would we ever want to explicitly NOT set the `DATE` field when we know there
is a valid time available?

So I've removed `HTTP_SERVER_EXPOSE_DATE` and instead set the DATE field only if the clock has been set.

Also, it seems unsafe to assume whether local or UTC time is being used when setting the clock
so have removed the default to force applications to be explicit.