Fix keyboard thread race condition.
Small test applications can execute and finish before the keyboard thread has started running.
Adding a small delay and an additional check handles this situation and avoids calling the keyb_XX functions.

This doesn't guarantee that `keyb_raw()` isn't called during cleanup as the race condition still exists.
This isn't a problem in itself, except for the call to `atexit()` which results in undefined behaviour
and in practice, a deadlock. It's not required as the keyboard thread calls `keyb_restore()` when it's done.