Fix lockup with spaces in paths
Whilst source directories must not contain spaces, there could be other directories with spaces in their names.
These must be detected to prevent infinite recusion with `make cs` and `make cs-dev`.

Also fix logic so that `CS_DIRS` is only expanded when used. This caused the same lockup to occur when
_any_ target in Makefile is invoked (e.g. dist-clean).