This change tries to address the following possible issue with servers that don't support keep-alive connections.
Initial conditions
------------------
3 consecutive GET requests + 1 POST request go to the same HTTP server.

Before this change
------------------
The HTTP client starts using HTTP pipelining (https://en.wikipedia.org/wiki/HTTP_pipelining) without making sure that the HTTP server is supporting it.
It will try to send the first 3 HTTP requests in one go. If the HTTP server does not support pipelining it will either return bad request response or a normal response
with Connection: close header. The first request will be successful but then the connection will be closed and the remaining two requests will be moved back to the waiting queue.
They will be AFTER the POST request which is a **problem**.

Once the POST request has completed the two GET requests will be moved back and forth between the waiting and execution queue which is another **problem**.

After this change
-----------------
The first GET request will be sent but the others will wait in the waiting queue.
Once the request has completed the Connection response header will be analyzed. 

a) If the header is Connection:close then the connection will be closed and there will be no need to move something from the execution queue back to the waiting queue
as there will be 0 requests in execution and the rest will be in the waiting queue.

b) if there is no Connection:close header in the response the HTTP client will assume that the HTTP server supports pipelining and will allow pipelining consecutive requests
 in the same Http Client Connection. This way the remaining 2 GET requests will be pipelined and the POST request will be executed at the right time.