FWFS stores objectIDs directly (breaking change) (#2321)
This PR simplifies FWFS further by storing object identifiers as image offsets. Advantages:

- No cache required
- Faster because we can jump directly to the referenced record
- Debugging and introspection is much easier

The references are stored as packed integers (1 - 4 bytes).

Note: Existing images need to be rebuilt. This change should have been included in #2315 as the use of numeric object IDs was part of the abstraction for object stores.