Add Host stubs for RTC mem read/write (#2588)
This PR adds stubs for the `system_rtc_mem_read` and `system_rtc_mem_write` functions so that code compiles for the `Host` arch.

No immediate use-cases fo this, but for future consideration it might be useful to make these functional using a backing file, similar to how flash memory is handled. One area where this might be useful is improving `System.restart()` behaviour, perhaps by forking a separate process which can be restarted.