Initial RP2040 networking support (#2607)
This PR adds networking support for the Pico-W via the Pico SDK, which uses the CYW43439 WiFi/Bluetooth chip driver https://github.com/georgerobotics/cyw43-driver plus LWIP.

The chip requires a firmware BLOB to be uploaded, which is about 240KBytes. The driver is patched to support firmware image compression. By default, the firmware is linked into the application image but may optionally be stored in a separate partition. The Basic_IFS sample demonstrates this.

One issue which appears to affect the ESP32 also is that attempting to scan for WiFi stations whilst a station connect is in progress will cause the scan to fail. The `HttpServer_ConfigNetwork` sample has been revised to detect this and retry, plus periodically refreshing the station list. Also, as the station list shows only the SSID (not the AP BSSID) it is confusing where multiple access points exist for a single SSID. The application now shows only a single entry for the AP with the strongest signal. Note: This de-duplication is handled by the application (not the Network Component) since another approach is to show all related APs for each SSID and allow the user to select a specific AP. This is not currently demonstrated in any samples.

Power management and sleep functions have not yet been implemented.