Fix Esp32 flashing (#2613)
This PR fixes a couple of problems related to ESP32 bootloader flashing.

The `make sdk-configclean` build target which currently does not full clear the SDK configuration information as expected.

The bootloader debug log may show the `SHA-256 comparison failed` message.
The reason for this is that, by default, esptool will auto-detect flash settings and overwrite those values in the bootloader header. Those settings are the mode, speed and flash size settings.

This behaviour is now disabled (by setting CONFIG_ESPTOOLPY_FLASHSIZE_DETECT=n),
and the user-configured parameter values passed to the SDK.

This also fixes the problem for later versions of esptool (4+) which ignores auto-detect unless bootloader checksumming is also disabled. This results in boot failure as the flash chip size is usually too small (default is 2MB).