Add USB library (#2623)
This PR adds a new library which aims to simplify the use of TinyUSB with Sming.

Note that TinyUSB doesn't currently support host mode for esp32s2/s3 so currently only works for Rp2040. The Rp2040 SDK is updated to the latest develop branch.

Check out the documentation/samples for further details.