Second (and final) attempt at fixing this issue.
Whilst using a function to 'hide' the stack return appears to work, integration tests fail in a couple of scenarios (FlashString, RapidXML) because of the way stack space gets used.
For Host builds, the most pragmatic solution is to make `_F(s)` return `s` directly.
That means it gives a `const char*` rather than `char*`, and any attempt to pass this to a function requiring a regular `char*` will fail.