Improve `Vector` and `HttpHeaders` iteration (#2745)
This PR improves use of C++ iterator pattern instead of traditional `for(i=0; i<list.count(); ++i)` approach.

**Template Vector `indexOf` methods**

Allows comparison with things other than an object instance, e.g. its name, depending on operator overloads.
Instead of recoding a comparison loop, just implement the appropriate `operator==` for the classes and use `Vector::indexOf`.

**Enable iteration for HttpHeaders**

This is based on HashMap but stores an enumeration for the key as it's more efficient.
HashMap iteration returns a proxy object, so here we provide a customised version with an implicit `String()` operator.
This makes for simpler usage.