Fixes to hardware config editing/display (#2765)
**If name not specified when loading config then 'save' raises exception**

For example, `Basic_Ota` sample doesn't have its `name` field set in `ota.hw`:

- Run `make hwconfig-edit`
- Click 'save'

So ensure we provide a default name based on filename.


**Ensure all sample .hw files have names**

For consistency.


**Remove 'unused_before' and 'unused_after' fields from `make hwconfig` output**

These are only used for map building so shouldn't be visible.
Example:

```
        "phy_init": {
            "device": "spiFlash",
            "address": "0x000fc000",
            "size": "4K",
            "type": "data",
            "subtype": "phy",
            "readonly": false,
            "encrypted": false,
            "filename": "$(FLASH_INIT_DATA)",
            "unused_before": 0, ///<
            "unused_after": 0   ///< Don't show
        },
```


**Don't set name, comment fields from option data**

For example, `Basic_Blink` sample uses default config. Running `make hwconfig` gives:

```
{
    "name": "Standard config with single ROM",
    "comment": "Should work with any Esp8266 variant",
    "arch": "Esp8266",
    "options": [],
```

If we run `make hwconfig HWCONFIG_OPTS=4m` we get this:

```
{
    "name": "",
    "comment": "",
    "arch": "Esp8266",
    "options": [
        "4m"
    ],
```

The `name` and `comment` have been wiped out, should be as before.


**Fix use of `not in` and `is not` operators**

e.g. `x not in [...]` preferred to `not x in [...]`